refactor(ir): remove `ibis.expr.streaming`
